requirements.txt
mss>=9.0
pyautogui>=0.9
easyocr>=1.7
paddleocr>=2.7
pytesseract>=0.3
Pillow>=10.0
PyQt5>=5.15
keyboard>=0.13
pynput>=1.7
ollama>=0.2
openai>=1.0
pyyaml>=6.0
keyring>=24.0
hypothesis>=6.0
pytest>=8.0
